feature: 研发商店新版设计 issue #1896 - 工作台相关页面新版设计
