+++ src/frontend/devops-atomstore/src/components/common/card/empty.vue
@@ -2,7 +2,7 @@
     <h3 class="atom-card empty-card">
         <img class="empty-logo" :src="image" />
         <p class="empty-prompt">{{emptyData.str}}
-            <router-link class="router-list" :to="{ name: 'workList', params: { type: emptyData.type } }">{{emptyData.tip}}</router-link>
+            <router-link class="router-list" :to="{ name: `${emptyData.type}Work` }">{{emptyData.tip}}</router-link>
         </p>
     </h3>
 </template>
