+++ src/frontend/devops-atomstore/src/views/edit_image.vue
@@ -460,10 +460,7 @@
 
             toImageList () {
                 this.$router.push({
-                    name: 'workList',
-                    params: {
-                        type: 'image'
-                    }
+                    name: 'imageWork'
                 })
             },
