+++ src/frontend/devops-atomstore/src/views/manage/index.vue
@@ -1,32 +1,22 @@
 <template>
-    <article class="store-manage">
-        <header class="manage-title">
+    <article class="g-store-main">
+        <header class="g-store-title">
             <span class="banner-des quick-route" @click="goToStore"> {{ $t('store.研发商店') }} </span>
             <i class="right-arrow banner-arrow"></i>
-            <span class="banner-des quick-route" @click="goToWorkList"> {{ $t('store.工作台') }}</span>
-            <i class="right-arrow banner-arrow"></i>
-            <span class="banner-des quick-route" @click="goToDetail">{{ type | typeFilter }}</span>
+            <span class="banner-des quick-route" @click="goToWorkList"> {{ $t('store.工作台') }}：{{ type | typeFilter }}</span>
             <i class="right-arrow banner-arrow"></i>
             <span class="banner-des">{{ $route.params.code }}</span>
         </header>
 
-        <bk-tab :active.sync="activeTab" type="unborder-card" class="manage-tabs" @tab-change="tabChange">
-            <bk-tab-panel v-for="(panel, index) in panels" v-bind="panel" :key="index">
-                <transition name="fade">
-                    <ul v-if="activeTab === panel.name && panel.showChildTab" class="manage-child-tabs">
-                        <li v-for="childPanel in panel.children.filter(x => !x.hidden)"
-                            :key="childPanel.name"
-                            @click="tabChange(childPanel.name)"
-                            :class="['manage-child-tab', { active: activeChildTab === childPanel.name }]"
-                        >{{ childPanel.label }}</li>
-                    </ul>
-                </transition>
-            </bk-tab-panel>
-        </bk-tab>
+        <transition-tab :panels="panels"
+            :transition-name.sync="transitionName"
+            @tab-change="tabChange"
+            @child-tab-change="childTabChange"
+        ></transition-tab>
 
-        <main v-bkloading="{ isLoading }" class="manage-main">
+        <main v-bkloading="{ isLoading }" class="g-store-body">
             <transition :name="transitionName">
-                <router-view v-if="Object.keys(detail).length > 0 && !isLoading" class="manage-route" v-bind="routekey"></router-view>
+                <router-view v-if="Object.keys(detail).length > 0 && !isLoading" class="g-store-route" v-bind="routekey"></router-view>
             </transition>
         </main>
     </article>
@@ -34,8 +24,13 @@
 
 <script>
     import { mapGetters } from 'vuex'
+    import transitionTab from '@/components/transition-tab.vue'
 
     export default {
+        components: {
+            transitionTab
+        },
+
         filters: {
             typeFilter (val) {
                 const bkLocale = window.devops || {}
@@ -63,6 +58,7 @@
                 isLoading: true,
                 type: '',
                 transitionName: '',
+                routekey: {},
                 panelMap: {
                     atom: [
                         { label: this.$t('store.概览'), name: 'overView' },
@@ -98,34 +94,7 @@
         computed: {
             ...mapGetters('store', {
                 'detail': 'getDetail'
-            }),
-
-            routekey () {
-                const res = {}
-                if (this.activeChildTab) res.key = this.activeChildTab
-                return res
-            }
-        },
-
-        watch: {
-            '$route.name' (val, oldVal) {
-                const calcIndex = (name) => {
-                    let res = ''
-                    this.panels.forEach((panel, index) => {
-                        if (name === panel.name) res = index + '0'
-                        if (panel.children) {
-                            panel.children.forEach((childPanel, childIndex) => {
-                                if (!panel.showChildTab) childIndex = 1
-                                if (name === childPanel.name) res = `${index}${childIndex}`
-                            })
-                        }
-                    })
-                    return +res
-                }
-                const diff = calcIndex(val) - calcIndex(oldVal)
-                this.transitionName = diff > 0 ? 'g-slide-left' : diff === 0 ? 'atom-fade' : 'g-slide-right'
-                this.calcActiveTab()
-            }
+            })
         },
 
         created () {
@@ -138,21 +107,19 @@
                 const params = this.$route.params || {}
                 this.type = params.type
                 this.panels = this.panelMap[this.type]
-                this.calcActiveTab()
             },
 
             tabChange (tabName) {
-                const currentPanel = this.panels.find((panel) => (panel.name === tabName || (panel.children && panel.children.some(x => x.name === tabName)))) || {}
-                const panelChildren = currentPanel.children
-                const name = panelChildren && tabName === currentPanel.name ? (panelChildren[0] || {}).name : tabName
+                this.routekey.key = undefined
+                const currentPanel = this.panels.find((panel) => (panel.name === tabName)) || {}
+                const panelChildren = currentPanel.children || []
+                const name = panelChildren.length ? (panelChildren[0] || {}).name : currentPanel.name
                 this.$router.push({ name })
             },
 
-            calcActiveTab (val) {
-                const name = val || this.$route.name
-                const currentPanel = this.panels.find((panel) => (panel.name === name || (panel.children && panel.children.some(x => x.name === name)))) || {}
-                this.activeTab = currentPanel.name
-                this.activeChildTab = currentPanel.name !== name && name
+            childTabChange (name) {
+                this.routekey.key = name
+                this.$router.push({ name })
             },
 
             initData () {
@@ -193,125 +160,15 @@
             },
 
             goToWorkList () {
+                const name = `${this.type}Work`
                 this.$router.push({
-                    name: 'workList',
-                    params: {
-                        type: this.type
-                    }
+                    name
                 })
             },
 
             goToStore () {
                 this.$router.push({ name: 'atomHome' })
-            },
-
-            goToDetail () {
-                let defaultPage = this.panels[0]
-                if (defaultPage.children && defaultPage.children.length > 0) defaultPage = defaultPage.children[0]
-                this.$router.push({ name: defaultPage.name })
             }
         }
     }
 </script>
-
-<style lang="scss" scoped>
-    .store-manage {
-        background-color: #f1f2f3;
-        height: 100vh;
-        color: #222;
-        display: flex;
-        max-height: 100vh;
-        padding-bottom: 40px;
-        flex-direction: column;
-        overflow: hidden;
-        .manage-title {
-            height: 56px;
-            line-height: 56px;
-            display: flex;
-            align-items: center;
-            padding-left: .32rem;
-            margin-bottom: 32px;
-            background-color: #fff;
-            color: #999;
-            .right-arrow {
-                height: 56px;
-                &::after {
-                    margin-top: -2.5px;
-                    border-color: #252935;
-                }
-            }
-            .quick-route {
-                color: #222;
-                cursor: pointer;
-            }
-        }
-        .manage-tabs {
-            width: 14.6rem;
-            margin: 0 auto;
-            box-shadow: 1px 2px 3px 0 rgba(0,0,0,0.05);
-            .manage-child-tabs {
-                height: 47px;
-                line-height: 47px;
-                padding: 12px 16px;
-                background-color: #fff;
-                &::after {
-                    content: '';
-                    display: table;
-                    clear: both;
-                }
-                .manage-child-tab {
-                    float: left;
-                    padding: 0 16px;
-                    font-size: 16px;
-                    line-height: 22px;
-                    color: #666;
-                    cursor: pointer;
-                    &.active {
-                        color: #1a6df3;
-                    }
-                    &:not(:last-child) {
-                        border-right: 1px solid #ebedf0;
-                    }
-                }
-            }
-            /deep/ .bk-tab-header {
-                background-color: #fff;
-                height: 64px;
-                line-height: 64px;
-                background-image: linear-gradient(transparent 63px,#dcdee5 0);
-                .bk-tab-label-list {
-                    height: 64px;
-                    .bk-tab-label-item {
-                        line-height: 64px;
-                        color: #666;
-                        &::after {
-                            height: 3px;
-                            width: 64px;
-                            left: 18px;
-                        }
-                        &.active {
-                            color: #3a84ff;
-                        }
-                        .bk-tab-label {
-                            font-size: 16px;
-                        }
-                    }
-                }
-            }
-            /deep/ .bk-tab-section {
-                padding: 0;
-            }
-        }
-        .manage-main {
-            width: 14.6rem;
-            margin: 16px auto 0;
-            position: relative;
-            flex: 1;
-            height: 0;
-            .manage-route {
-                height: 100%;
-                box-shadow: 1px 2px 3px 0 rgba(0,0,0,0.05);
-            }
-        }
-    }
-</style>
