+++ src/frontend/devops-atomstore/src/views/release_progress.vue
@@ -261,10 +261,7 @@
 
             toAtomList () {
                 this.$router.push({
-                    name: 'workList',
-                    params: {
-                        type: 'atom'
-                    }
+                    name: 'atomWork'
                 })
             },
             toAtomStore () {
