+++ src/frontend/devops-atomstore/src/views/upgrade_template.vue
@@ -289,10 +289,7 @@
             },
             toAtomList () {
                 this.$router.push({
-                    name: 'workList',
-                    params: {
-                        type: 'template'
-                    }
+                    name: 'templateWork'
                 })
             },
             toAtomStore (val) {
