+++ src/frontend/devops-atomstore/src/views/workList/atom.vue
@@ -10,65 +10,70 @@
                 <i class="devops-icon icon-close-circle-shape clear-icon" v-else @click="clearSearch"></i>
             </section>
         </div>
-        <bk-table style="margin-top: 15px;" :empty-text="$t('store.暂时没有插件')"
-            :data="renderList"
-            :pagination="pagination"
-            @page-change="pageChanged"
-            @page-limit-change="pageCountChanged"
-            v-bkloading="{ isLoading }"
-        >
-            <bk-table-column :label="$t('store.插件名称')">
-                <template slot-scope="props">
-                    <span class="atom-name" :title="props.row.name" @click="routerAtoms(props.row.atomCode)">{{ props.row.name }}</span>
-                </template>
-            </bk-table-column>
-            <bk-table-column :label="$t('store.调试项目')" prop="projectName"></bk-table-column>
-            <bk-table-column :label="$t('store.开发语言')" prop="language"></bk-table-column>
-            <bk-table-column :label="$t('store.版本')" prop="version"></bk-table-column>
-            <bk-table-column :label="$t('store.状态')">
-                <template slot-scope="props">
-                    <div class="bk-spin-loading bk-spin-loading-mini bk-spin-loading-primary" v-if="props.row.atomStatus === 'COMMITTING' || props.row.atomStatus === 'BUILDING' || props.row.atomStatus === 'BUILD_FAIL' || props.row.atomStatus === 'TESTING' || props.row.atomStatus === 'AUDITING' || props.row.atomStatus === 'UNDERCARRIAGING'">
-                        <div class="rotate rotate1"></div>
-                        <div class="rotate rotate2"></div>
-                        <div class="rotate rotate3"></div>
-                        <div class="rotate rotate4"></div>
-                        <div class="rotate rotate5"></div>
-                        <div class="rotate rotate6"></div>
-                        <div class="rotate rotate7"></div>
-                        <div class="rotate rotate8"></div>
-                    </div>
-                    <span class="atom-status-icon success" v-if="props.row.atomStatus === 'RELEASED'"></span>
-                    <span class="atom-status-icon fail" v-if="props.row.atomStatus === 'GROUNDING_SUSPENSION'"></span>
-                    <span class="atom-status-icon obtained" v-if="props.row.atomStatus === 'AUDIT_REJECT' || props.row.atomStatus === 'UNDERCARRIAGED'"></span>
-                    <span class="atom-status-icon devops-icon icon-initialize" v-if="props.row.atomStatus === 'INIT'"></span>
-                    <span>{{ $t(atomStatusList[props.row.atomStatus]) }}</span>
-                </template>
-            </bk-table-column>
-            <bk-table-column :label="$t('store.修改人')" prop="modifier"></bk-table-column>
-            <bk-table-column :label="$t('store.修改时间')" prop="updateTime" width="150"></bk-table-column>
-            <bk-table-column :label="$t('store.操作')" width="240" class-name="handler-btn">
-                <template slot-scope="props">
-                    <span class="upgrade-btn"
-                        v-if="props.row.atomStatus === 'GROUNDING_SUSPENSION' || props.row.atomStatus === 'AUDIT_REJECT' || props.row.atomStatus === 'RELEASED'"
-                        @click="editHandle('upgradeAtom', props.row.atomId)"> {{ $t('store.升级') }} </span>
-                    <span class="install-btn"
-                        v-if="props.row.atomStatus === 'RELEASED'"
-                        @click="installAHandle(props.row.atomCode)"> {{ $t('store.安装') }} </span>
-                    <span class="shelf-btn"
-                        v-if="props.row.atomStatus === 'INIT' || props.row.atomStatus === 'UNDERCARRIAGED'"
-                        @click="editHandle('shelfAtom', props.row.atomId)"> {{ $t('store.上架') }} </span>
-                    <span class="obtained-btn"
-                        v-if="props.row.atomStatus === 'AUDIT_REJECT' || props.row.atomStatus === 'RELEASED' || (props.row.atomStatus === 'GROUNDING_SUSPENSION' && props.row.releaseFlag)"
-                        @click="offline(props.row)"> {{ $t('store.下架') }} </span>
-                    <span class="schedule-btn"
-                        v-if="props.row.atomStatus === 'COMMITTING' || props.row.atomStatus === 'BUILDING' || props.row.atomStatus === 'BUILD_FAIL'
-                            || props.row.atomStatus === 'TESTING' || props.row.atomStatus === 'AUDITING'"
-                        @click="routerProgress(props.row)"> {{ $t('store.进度') }} </span>
-                    <span class="delete-btn" v-if="!props.row.releaseFlag" @click="deleteAtom(props.row)"> {{ $t('store.删除') }} </span>
-                </template>
-            </bk-table-column>
-        </bk-table>
-
+        <main class="g-scroll-pagination-table">
+            <bk-table style="margin-top: 15px;"
+                :empty-text="$t('store.暂时没有插件')"
+                :outer-border="false"
+                :header-border="false"
+                :header-cell-style="{ background: '#fff' }"
+                :data="renderList"
+                :pagination="pagination"
+                @page-change="pageChanged"
+                @page-limit-change="pageCountChanged"
+                v-bkloading="{ isLoading }"
+            >
+                <bk-table-column :label="$t('store.插件名称')">
+                    <template slot-scope="props">
+                        <span class="atom-name" :title="props.row.name" @click="routerAtoms(props.row.atomCode)">{{ props.row.name }}</span>
+                    </template>
+                </bk-table-column>
+                <bk-table-column :label="$t('store.调试项目')" prop="projectName"></bk-table-column>
+                <bk-table-column :label="$t('store.开发语言')" prop="language"></bk-table-column>
+                <bk-table-column :label="$t('store.版本')" prop="version"></bk-table-column>
+                <bk-table-column :label="$t('store.状态')">
+                    <template slot-scope="props">
+                        <div class="bk-spin-loading bk-spin-loading-mini bk-spin-loading-primary" v-if="props.row.atomStatus === 'COMMITTING' || props.row.atomStatus === 'BUILDING' || props.row.atomStatus === 'BUILD_FAIL' || props.row.atomStatus === 'TESTING' || props.row.atomStatus === 'AUDITING' || props.row.atomStatus === 'UNDERCARRIAGING'">
+                            <div class="rotate rotate1"></div>
+                            <div class="rotate rotate2"></div>
+                            <div class="rotate rotate3"></div>
+                            <div class="rotate rotate4"></div>
+                            <div class="rotate rotate5"></div>
+                            <div class="rotate rotate6"></div>
+                            <div class="rotate rotate7"></div>
+                            <div class="rotate rotate8"></div>
+                        </div>
+                        <span class="atom-status-icon success" v-if="props.row.atomStatus === 'RELEASED'"></span>
+                        <span class="atom-status-icon fail" v-if="props.row.atomStatus === 'GROUNDING_SUSPENSION'"></span>
+                        <span class="atom-status-icon obtained" v-if="props.row.atomStatus === 'AUDIT_REJECT' || props.row.atomStatus === 'UNDERCARRIAGED'"></span>
+                        <span class="atom-status-icon devops-icon icon-initialize" v-if="props.row.atomStatus === 'INIT'"></span>
+                        <span>{{ $t(atomStatusList[props.row.atomStatus]) }}</span>
+                    </template>
+                </bk-table-column>
+                <bk-table-column :label="$t('store.修改人')" prop="modifier"></bk-table-column>
+                <bk-table-column :label="$t('store.修改时间')" prop="updateTime" width="150"></bk-table-column>
+                <bk-table-column :label="$t('store.操作')" width="240" class-name="handler-btn">
+                    <template slot-scope="props">
+                        <span class="upgrade-btn"
+                            v-if="props.row.atomStatus === 'GROUNDING_SUSPENSION' || props.row.atomStatus === 'AUDIT_REJECT' || props.row.atomStatus === 'RELEASED'"
+                            @click="editHandle('upgradeAtom', props.row.atomId)"> {{ $t('store.升级') }} </span>
+                        <span class="install-btn"
+                            v-if="props.row.atomStatus === 'RELEASED'"
+                            @click="installAHandle(props.row.atomCode)"> {{ $t('store.安装') }} </span>
+                        <span class="shelf-btn"
+                            v-if="props.row.atomStatus === 'INIT' || props.row.atomStatus === 'UNDERCARRIAGED'"
+                            @click="editHandle('shelfAtom', props.row.atomId)"> {{ $t('store.上架') }} </span>
+                        <span class="obtained-btn"
+                            v-if="props.row.atomStatus === 'AUDIT_REJECT' || props.row.atomStatus === 'RELEASED' || (props.row.atomStatus === 'GROUNDING_SUSPENSION' && props.row.releaseFlag)"
+                            @click="offline(props.row)"> {{ $t('store.下架') }} </span>
+                        <span class="schedule-btn"
+                            v-if="props.row.atomStatus === 'COMMITTING' || props.row.atomStatus === 'BUILDING' || props.row.atomStatus === 'BUILD_FAIL'
+                                || props.row.atomStatus === 'TESTING' || props.row.atomStatus === 'AUDITING'"
+                            @click="routerProgress(props.row)"> {{ $t('store.进度') }} </span>
+                        <span class="delete-btn" v-if="!props.row.releaseFlag" @click="deleteAtom(props.row)"> {{ $t('store.删除') }} </span>
+                    </template>
+                </bk-table-column>
+            </bk-table>
+        </main>
         <template v-if="createAtomsideConfig.show">
             <bk-sideslider
                 class="create-atom-slider g-slide-radio"
@@ -535,7 +540,7 @@
                     query: {
                         code,
                         type: 'atom',
-                        from: 'workList'
+                        from: 'atomWork'
                     }
                 })
             },
